Added new buttons
Now the calculator has buttons for brackets, deletion, square root.
Staples:
Now it is possible to write equations from brackets
Deletion:
It is now possible to erase unnecessary entries. But at the moment it works a little crooked and after the first deletion you need to click 2 times to delete subsequent characters.
Square root:
While this function does not work